MAJ 30/06/2014
- Passage des MDP en sha1
- Requêtes préparés
- Revue de la structure du code sur les pages (ajout d'include pour le
code rémanent)
- Optimisation du module de recherche
- Sécurisation de l'affichage d'un joueur (si il n'existe pas alors une
erreur pré faite)
- Rajout du rôle guest (pour les joueurs du serveur, tout est visible
sauf les comptes)

+++ content/delete_player.php
@@ -2,7 +2,7 @@
 if(!Auth::isLogged() && !Auth::isAdmin()){
 	header('Location:'.WEBROOT);
 }
-if(isset($_GET['j'])) {
+if(isset($_GET['j']) && Auth::isAdmin()) {
 	$j = $_GET['j'];
 	$DB->exec("DELETE FROM players WHERE playerid='$j'");
 	$DB->exec("DELETE FROM vehicles WHERE pid='$j'");
